fixed the folder gituationship
