+++ java/datastructures/lib/bin/main/datastructures/linkedlist/LinkedList.class
