+++ java/datastructures/lib/src/main/java/codechallenges/linkedlist/LinkedListChallenges.java
