+++ java/datastructures/lib/src/main/java/codechallenges/linkedlist/Node.java
