+++ java/datastructures/lib/src/main/java/datastructures/linkedlist/Node.java
