+++ java/datastructures/lib/src/main/java/datastructures/queue/Library.java
@@ -1,0 +1,10 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package datastructures.queue;
+
+public class Library {
+    public boolean someLibraryMethod() {
+        return true;
+    }
+}
