+++ java/datastructures/lib/src/main/java/datastructures/trees/Library.java
@@ -1,0 +1,12 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package datastructures.trees;
+
+public class Library {
+    public boolean someLibraryMethod() {
+        return true;
+    }
+
+
+}
