+++ java/datastructures/lib/src/main/java/datastructures/trees/Node.java
@@ -1,0 +1,46 @@
+package datastructures.trees;
+
+public class Node {
+  private int value;
+  private Node left;
+  private Node right;
+
+  public Node(int value) {
+    this.value = value;
+    this.left = null;
+    this.right = null;
+  }
+
+  public int getValue() {
+    return value;
+  }
+
+  public void setValue(int value) {
+    this.value = value;
+  }
+
+  public Node getLeft() {
+    return left;
+  }
+
+  public void setLeft(Node left) {
+    this.left = left;
+  }
+
+  public Node getRight() {
+    return right;
+  }
+
+  public void setRight(Node right) {
+    this.right = right;
+  }
+
+  @Override
+  public String toString() {
+    return "Node{" +
+      "value=" + value +
+      ", left=" + left +
+      ", right=" + right +
+      '}';
+  }
+}
