+++ java/datastructures/lib/src/test/java/datastructures/queue/LibraryTest.java
@@ -1,0 +1,11 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package datastructures.queue;
+
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.*;
+
+class LibraryTest {
+
+}
