+++ java/datastructures/lib/src/test/java/datastructures/stack/LibraryTest.java
@@ -1,0 +1,11 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package datastructures.stack;
+
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.*;
+
+class LibraryTest {
+
+}
