--- java/datastructures/linkedlist/README.md
@@ -1,17 +1,0 @@
-# Data Structures
-
-## Linked List Insertions
-
-Take an existing Linked List program and insert values at the end, before, or after a particular node.
-
-### Approach & Efficiency
-
-I modeled my code and tests behind what already existed from the previous code challenge.
-
-### Solution
-
-To run the code, create a new linked list and add values to value.
-
-Location: `codechallenges/linkedlist/LinkedListChallenges`
-
-Method: `linkedListChallenge06()`
