--- java/datastructures/queue/.gitignore
@@ -1,8 +1,0 @@
-# Ignore Gradle project-specific cache directory
-.gradle
-
-# Ignore Gradle build output directory
-build
-
-# Ignore .idea folder
-.idea
