--- java/datastructures/queue/lib/src/main/java/queue/Library.java
@@ -1,10 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package queue;
-
-public class Library {
-    public boolean someLibraryMethod() {
-        return true;
-    }
-}
