--- java/datastructures/stack/lib/src/main/java/stack/Library.java
@@ -1,10 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package stack;
-
-public class Library {
-    public boolean someLibraryMethod() {
-        return true;
-    }
-}
