--- java/datastructures/trees/lib/src/main/java/trees/Library.java
@@ -1,12 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package trees;
-
-public class Library {
-    public boolean someLibraryMethod() {
-        return true;
-    }
-
-
-}
